#324 : conditional boxing

+++ elenasrc3/elenasm/vmparser.cpp
@@ -63,7 +63,7 @@ int VMTapeParser :: writeBuildScriptArgumentList(ScriptEngineReaderBase& reader,
 
    pos_t position = writer.writeAndGetArgPosition(VM_ALLOC_CMD, 0);
 
-   size_t counter = 0;
+   int counter = 0;
    ScriptBookmark bm = callStack.pop();
    while (!bm.compare(terminator)) {
       writeBuildScriptStatement(reader, bm, callStack, writer);
@@ -77,7 +77,7 @@ int VMTapeParser :: writeBuildScriptArgumentList(ScriptEngineReaderBase& reader,
 
    // NOTE : we need to reverse the order of arguments
    if (counter > 1) {
-      for (size_t i = 0; i < counter; i++) {
+      for (int i = 0; i < counter; i++) {
          writer.fixArg(argPositions[i], counter - i - 1);
       }
    }
